move LocalParticle to src/particle.rs
