+++ src/physics/gravity/tests/parallel.rs
@@ -14,10 +14,10 @@ use crate::mass;
 use crate::physics::gravity::plugin::GravityPlugin;
 use crate::physics::gravity::GravityParameters;
 use crate::physics::gravity::Solver;
-use crate::physics::LocalParticle;
 use crate::physics::PhysicsPlugin;
 use crate::physics::Timestep;
 use crate::position::Position;
+use crate::prelude::LocalParticle;
 use crate::simulation::Simulation;
 use crate::test_utils::run_system_on_sim;
 use crate::units::Mass;
