+++ src/physics/hydrodynamics/mod.rs
@@ -9,7 +9,6 @@ pub use self::parameters::HydrodynamicsParameters;
 use self::quadtree::construct_quad_tree_system;
 use self::quadtree::get_particles_in_radius;
 use self::quadtree::QuadTree;
-use super::LocalParticle;
 use super::Timestep;
 use crate::density;
 use crate::domain::extent::Extent;
@@ -18,6 +17,7 @@ use crate::mass::Mass;
 use crate::named::Named;
 use crate::performance_parameters::PerformanceParameters;
 use crate::position::Position;
+use crate::prelude::LocalParticle;
 use crate::pressure;
 use crate::quadtree::LeafDataType;
 use crate::simulation::RaxiomPlugin;
