+++ src/prelude.rs
@@ -2,9 +2,9 @@ pub use crate::communication::WorldRank;
 pub use crate::communication::WorldSize;
 pub use crate::mass::Mass;
 pub use crate::named::*;
+pub use crate::particle::LocalParticle;
 pub use crate::physics::hydrodynamics::HydrodynamicsPlugin;
 pub use crate::physics::GravityPlugin;
-pub use crate::physics::LocalParticle;
 pub use crate::physics::Timestep;
 pub use crate::position::Position;
 pub use crate::simulation::Simulation;
