+++ src/visualization/mod.rs
@@ -27,9 +27,9 @@ use crate::communication::Rank;
 use crate::domain::determine_global_extent_system;
 use crate::named::Named;
 use crate::physics::gravity;
-use crate::physics::LocalParticle;
 use crate::physics::StopSimulationEvent;
 use crate::position::Position;
+use crate::prelude::LocalParticle;
 use crate::quadtree::QuadTreeVisualizationPlugin;
 use crate::simulation::RaxiomPlugin;
 use crate::simulation::Simulation;
